Fix error

+++ lib/crypto/crypto_key_value.dart
@@ -17,21 +17,21 @@ class CryptoKeyValue extends HiveKeyValue {
   static const _nextSignedPreKeyId = 'next_signed_pre_key_id';
   static const _activeSignedPreKeyId = 'active_signed_pre_key_id';
 
-  int get localRegistrationId => box.get(_localRegistrationId, defaultValue: 0);
+  int get localRegistrationId => box.get(_localRegistrationId, defaultValue: 0) as int;
   set localRegistrationId(int registrationId) =>
       box.put(_localRegistrationId, registrationId);
 
   int get nextPreKeyId =>
-      box.get(_nextPreKeyId, defaultValue: generateRandomInt(maxValue));
+      box.get(_nextPreKeyId, defaultValue: generateRandomInt(maxValue)) as int;
   set nextPreKeyId(int preKeyId) => box.put(_nextPreKeyId, preKeyId);
 
   int get nextSignedPreKeyId =>
-      box.get(_nextSignedPreKeyId, defaultValue: generateRandomInt(maxValue));
+      box.get(_nextSignedPreKeyId, defaultValue: generateRandomInt(maxValue)) as int;
   set nextSignedPreKeyId(int preKeyId) =>
       box.put(_nextSignedPreKeyId, preKeyId);
 
   int get activeSignedPreKeyId =>
-      box.get(_activeSignedPreKeyId, defaultValue: -1);
+      box.get(_activeSignedPreKeyId, defaultValue: -1) as int;
   set activeSignedPreKeyId(int preKeyId) =>
       box.put(_activeSignedPreKeyId, preKeyId);
 }
